% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve}
\usage{
auc(p)
}
\arguments{
\item{p}{a numeric vector}
}
\value{
a scalar, which is the AUC
}
\description{
Calculates AUC under a probability curve
}
\examples{
p <- c(1,3,5,7,8,2)
auc(p)

}
