% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmooth.R
\name{gsmooth}
\alias{gsmooth}
\title{Grishin's profile smoothing}
\usage{
gsmooth(x, window.half = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{window.half}{Integer size of the half window}
}
\value{
a vector of same length as input
}
\description{
Smooths a profiles by averaging over windows, as implemented in AL2CO by Nick Grishin
}
\examples{
x = c(1,5,4,3,3,7,8)
gsmooth(x)

}
