% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kn.R
\name{Kn}
\alias{Kn}
\title{Effective number of states}
\usage{
Kn(p)
}
\arguments{
\item{p}{a numeric vector, not necessarily sum(p) = 1 (but will be renormalized as a probability  vector)}
}
\value{
a scalar representing effective number of states that contribute to p
}
\description{
GIven a distribution, it calculates the effective number of states (exp(entropy))
}
\examples{
p = c(1,1,0,0)
Kn(p)

}
