% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMnorm.R
\name{mMnorm}
\alias{mMnorm}
\title{min-max normalization}
\usage{
mMnorm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector (normalized)
}
\description{
Scale vector so that range is [0,1]
}
\examples{
x = c(1,1,0,0,.5,.8)
mMnorm(x)

}
