% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnorm.R
\name{mnorm}
\alias{mnorm}
\title{mean-normalization}
\usage{
mnorm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of same length as input
}
\description{
Normalizes elements of vector by dividing by mean
}
\examples{
x = c(1,5,4)
mnorm(x)

}
