% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offdiag.R
\name{offdiag}
\alias{offdiag}
\title{Off-diagonal part of a matrix}
\usage{
offdiag(A)
}
\arguments{
\item{A}{a square matrix}
}
\value{
a square off-diagonal matrix
}
\description{
Given a matrix A, returns the a matrix with off-diagonal elements and 0's in diagonal
}
\examples{
A = matrix(c(1,2,3,4), 2, 2)
offdiag(A)

}
