% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r9norm.R
\name{r9norm}
\alias{r9norm}
\title{Decile normalization}
\usage{
r9norm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
an integer vector of normalized values
}
\description{
rank-normalize using 9 quantiles (just as ConSurfDB des to colour 3D structures) this is further smoothing
}
\examples{

x = rnorm(100)
r9norm(x)

}
