% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm.R
\name{rank_norm}
\alias{rank_norm}
\title{Rank-normalize a vector}
\usage{
rank_norm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector with elements in range (0,1) proportional to ranks of x
}
\description{
Given vector x, return vector of ranks, min-max normalized
}
\examples{
x = c(1,1,0,0,.5,.4)
rnorm(x)

}
