% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znorm.R
\name{znorm}
\alias{znorm}
\title{z-score normalization}
\usage{
znorm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of z-scores
}
\description{
Normalize z = (x - mean(x)) / sd(x)
}
\examples{
x = rnorm(100)
znorm(x)

}
